修复场景Bug

+++ Server/Model/Module/Demo/Config/UnitConfig.cs
@@ -1,3 +1,5 @@
+using System;
+
 namespace ETModel
 {
 	[Config((int)(AppType.ClientH |  AppType.ClientM | AppType.Gate | AppType.Map))]
@@ -7,7 +9,7 @@ namespace ETModel
 
 	public class UnitConfig: IConfig
 	{
-		public long Id { get; set; }
+		public Int32 Id { get; set; }
 		public string Name;
 		public string Desc;
 		public int Position;
